Entities for WeatherApi are added

+++ app/src/main/java/com/example/innowiseweatherapplication/api/IWeatherService.kt
@@ -1,13 +1,14 @@
 package com.example.innowiseweatherapplication.api
 
 import com.example.innowiseweatherapplication.model.entity.WeatherClass
-import io.reactivex.rxjava3.core.Observable
 import retrofit2.http.GET
 import retrofit2.http.Query
+import io.reactivex.Observable
 
 interface IWeatherService {
 
     @GET("forecast?appid=${WeatherRetrofitApi.API_KEY}")
-    fun getWeather(@Query("lat") lat: Long,@Query("lon") lon: Long):Observable<WeatherClass>
+    fun getWeather(@Query("lat") lat: Double,@Query("lon") lon: Double): Observable<WeatherClass>
+
 
 }
